Added tags to Profile Page
