+++ js/src/app/user/[userId]/_components/UserProfile/ProfilePicture/ProfilePictureSkeleton.tsx
@@ -1,5 +1,5 @@
 import { Skeleton } from "@mantine/core";
 
 export default function ProfilePictureSkeleton() {
-  return <Skeleton height={32} width={200} mx={"md"} mb={"xs"} />;
+  return <Skeleton height={150} width={150} mx={"md"} mb={"xs"} />;
 }
