+++ js/src/app/user/[userId]/_components/UserProfile/UserProfileHeaderSkeleton.tsx
@@ -1,5 +1,14 @@
-import { Skeleton } from "@mantine/core";
+import { Flex, Skeleton, Stack } from "@mantine/core";
 
 export default function UserProfileHeaderSkeleton() {
-  return <Skeleton height={32} width={200} mx={"md"} mb={"xs"} />;
+  return (
+    <>
+      <Flex direction={"row"} wrap={"wrap"}>
+        <Stack justify="center">
+          <Skeleton height={30} width={190}/>
+          <Skeleton height={30} width={190}/>
+        </Stack>
+      </Flex>
+    </>
+  );
 }
